Add: randomizeIds option
The logic behind this option was always active before, but the result was wrong. The intention was to prevent ID collisions caused by multiple avatars on one page. This has also worked so far. - Unless an avatar was created with the same options. This bug has now been fixed.

However, the resulting XML with this corrected logic is no longer deterministic. Therefore there is now the additional option `randomizeIds` which activates the behavior only when needed.